pulling out add and remove buttons for workers
